requirements.txt
fastapi
uvicorn
SQLAlchemy
pydantic
python-jose
